.agent/skills/generating-leads/scripts/requirements.txt
instaloader
playwright
typer
beautifulsoup4
pandas
python-dotenv
